# packages/server/swagger.yaml
openapi: 3.0.2
info:
  title: Platform Core API
  version: v1
  description: |
    # Introduction
    The Platform Core API defines a set of capabilities that can be used to manage a tenant of the MATTR Platform.

    The API can be used to manage Decentralised Identifiers, issue Credentials directly or using OpenID Connect flows, and verify Messages signed with DIDs.

    ## API Reference Versioning
    A previous version of the API Reference can be [accessed here](/api-ref-previous).

    # Pagination
    Most list operations in the API use pagination that can be controlled by a cursor method using the `cursor` and `limit` query parameters.

    **Example on [Retrieve List of Credentials](#operation/retrieveListCreds)**
    ```
    GET https://tenant.vii.mattr.global/core/v1/credentials
    ?limit=100
    &cursor=Y3JlYXRlZEF0PTIwMjAtMTAtMDhUMjMlM0ExMyUzQTE3Ljg5NtZGUxZWEyNzQ4MWI4
    ```

    * The `nextCursor` is found at the start of each returned range of credential entries and identifies the last object in the list.
    * The `limit` determines how many entries are returned in that request, with a maximum value of 1000.

    Requesting a page after the last value in the list will return an empty `data` object.

    ``` json
    {
    "data": []
    }
    ```

    Not using a query parameter defaults the response to return the first range of credential entries with a limit of 100.


    # Authorization
...

    <SecurityDefinitions />
  contact:
...
security:
  - bearerAuth: []
paths:
...
  '/v1/dids/{id}':
    get:
      tags:
        - DIDs
      summary: Resolve a DID
      description: |-
        Resolves the DID provided in the path to its DID Document from the DID method and method identifier. The DID Document may contain cryptographic material like a public key and service endpoints.

        Example:

        `did:sov:1234556` -  will be referenced on the Sovrin network
        `did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9` - the public key is encapsulated in the DID

        #### Returns

        The DID Document is returned along with meta-data already held about the DID by the tenant;
        - Registration Status: Given there can be a lag in writing on-ledger DIDs this status will be updated over a short timeframe (usually within 10 minutes).
        - Registered: Timestamp of registration.
        - Keys: A reference to the Keys in a KMS
      responses:
        '200':
          description: A DID Document and meta-data
          content:
            application/json:
              schema:
                type: object
                properties:
                  didDocument:
                    type: object
                  registrationStatus:
                    $ref: '#/paths/~1v1~1dids/post/responses/201/content/application~1json/schema/properties/registrationStatus'
                  localMetadata:
                    $ref: '#/paths/~1v1~1dids/post/responses/201/content/application~1json/schema/properties/metadata'
              examples:
                'Example did:key':
                  value:
                    didDocument:
                      '@context': 'https://w3.org/ns/did/v1'
                      id: 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                      publicKey:
                        - id: 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                          type: Ed25519VerificationKey2018
                          controller: 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                          publicKeyBase58: 6Prt5jwEFEDtz1sPwQSdqNH6Nf5fy87dpPg23CGWPanm
                      authentication:
                        - 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                      assertionMethod:
                        - 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                      capabilityDelegation:
                        - 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                      capabilityInvocation:
                        - 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                      keyAgreement:
                        - id: 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6LSo9fse5qkJjkZTSHRP69mvEw381Tww3umcjXaXQ7Zz71G'
                          type: X25519KeyAgreementKey2019
                          controller: 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                          publicKeyBase58: CUVi7n2tDH2pN3uerSdpbeiZGrvqESjcjkou2wU3GjEW
                    registrationStatus: COMPLETED
                    localMetadata:
                      keys:
                        - kmsKeyId: 6Prt5jwEFEDtz1sPwQSdqNH6Nf5fy87dpPg23CGWPanm
                          didDocumentKeyId: 'did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9#z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9'
                      registered: 1588562204304
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/paths/~1v1~1dids/post/responses/400/content/application~1json/schema'
              examples:
                '404':
                  value:
                    code: NotFound
                    message: Resource Not Found
      operationId: resolveDid
      x-codeSamples:
        - lang: curl
          source: |
            curl --request GET \
              --url https://tenant.vii.mattr.global/core/v1/dids/did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9 \
              --header 'Accept: application/json' \
              --header 'Authorization: Bearer REPLACE_ACCESS_TOKEN'
    delete:
      tags:
        - DIDs
      summary: Delete a DID
      responses:
        '204':
          description: DID successfully deleted
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/paths/~1v1~1dids/post/responses/400/content/application~1json/schema'
              examples:
                '404':
                  value:
                    code: NotFound
                    message: Resource Not Found
      operationId: deleteDid
      x-codeSamples:
        - lang: curl
          source: |
            curl --request DELETE \
              --url https://tenant.vii.mattr.global/core/v1/dids/did:key:z6Mkjr7vfzBfamiN6Wi6cyQUgTq6CEMXP1MzWQawsUEXJoa9 \
              --header 'Accept: application/json' \
              --header 'Authorization: Bearer REPLACE_ACCESS_TOKEN'
    parameters:
      - name: id
        in: path
        description: Issuer ID
        schema:
          type: string
          format: uuid
          example: 41458e5a-9092-40b7-9a26-d4eb43c5792f
        required: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    clientRegistrationAuth:
      type: http
      scheme: bearer
      bearerFormat: Client registration_access_token
...
servers:
...
    variables:
...